+++ src/components/index.ts
@@ -1,4 +1,4 @@
-import VuetifyToolkit from './framework'
+import { VueConstructor, Component } from 'vue'
 import VTreeSelect from './VTreeSelect'
 import VCascader from './VCascader'
 import VDataGridSelect from './VDataGridSelect'
@@ -6,7 +6,31 @@ import VDateTimeSelect from './VDateTimeSelect'
 import VMdView from './VMdView'
 import VAdvDataTable from './VAdvDataTable'
 
-export default VuetifyToolkit
+export interface VuetifyToolkitUseOptions {
+  components?: Record<string, Component>
+}
+
+const defaultComponents = {
+  'v-tree-select': VTreeSelect,
+  'v-cascader': VCascader,
+  'v-data-grid-select': VDataGridSelect,
+  'v-date-time-select': VDateTimeSelect,
+  'v-md-view': VMdView,
+  'v-adv-data-table': VAdvDataTable
+}
+
+function install (v: VueConstructor, args?: VuetifyToolkitUseOptions): VueConstructor<Vue> {
+  const components = args ? args.components : defaultComponents
+  for (const key in components) {
+    const component = components[key]
+    if (component) {
+      v.component(key, component as typeof v)
+    }
+  }
+  return v
+}
+
+export default install
 
 export {
   VTreeSelect,
@@ -17,23 +41,6 @@ export {
   VAdvDataTable
 }
 
-const install = VuetifyToolkit.install
-
-const components = {
-  'v-tree-select': VTreeSelect,
-  'v-cascader': VCascader,
-  'v-data-grid-select': VDataGridSelect,
-  'v-date-time-select': VDateTimeSelect,
-  'v-md-view': VMdView,
-  'v-adv-data-table': VAdvDataTable
-}
-
-VuetifyToolkit.install = (Vue, args) => {
-  install.call(VuetifyToolkit, Vue, {
-    components
-  })
-}
-
 if (typeof window !== 'undefined' && window.Vue) {
-  window.Vue.use(VuetifyToolkit)
+  window.Vue.use(install)
 }
